fix: DRY types
