+++ src/Types/common.ts
@@ -1,10 +1,29 @@
 import { InputHTMLAttributes } from 'react';
 
-import { BookingData } from './API';
-
 export type Media = {
   url: string;
   alt: string;
+};
+
+export type BookingData = {
+  id: string;
+  dateFrom: string;
+  dateTo: string;
+  guest: number;
+  created: string;
+  updated: string;
+  venue: Venue;
+  customer: User;
+};
+
+export type ProfileData = {
+  name: string;
+  email: string;
+  bio: string;
+  avatar: Media;
+  banner: Media;
+  venueManager: boolean;
+  venues: Venue[];
 };
 
 export type Venue = {
@@ -22,22 +41,6 @@ export type Venue = {
   maxGuests: number;
   meta: Meta;
   bookings: BookingData[];
-};
-
-export type VenueCreate = {
-  name: string;
-  description: string;
-  price: number;
-  maxGuests: number;
-  rating: number;
-  media: Media[];
-  location: {
-    address: string;
-    city: string;
-    zip: string;
-    country: string;
-  };
-  meta: Meta;
 };
 
 export type Booking = {
